🚀 Added simple way to return files
Now u can go to `yourpage/files/file` and you can open it. [FILES ONLY IN SKRIPT/TEMPLATES]
Function is tested but if you want to improve it, **do it** :)
I added this due to #30

When i will add config you will have way to turn this on and off.